[NET-765] Refactor database module structure
