+++ cmdb/data_storage/connection.py
@@ -1,5 +1,5 @@
 # DATAGERRY - OpenSource Enterprise CMDB
-# Copyright (C) 2019 NETHINKS GmbH
+# Copyright (C) 2019 - 2020 NETHINKS GmbH
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -12,7 +12,7 @@
 # GNU Affero General Public License for more details.
 #
 # You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 """
 Database-Connection
@@ -26,7 +26,7 @@ from pymongo.database import Database
 from pymongo.errors import ConnectionFailure
 
 from cmdb.data_storage import CLIENT
-from cmdb.data_storage.errors import DatabaseConnectionError
+from cmdb.data_storage.errors.connection_errors import DatabaseConnectionError
 
 
 class ConnectionStatus:
