+++ cmdb/data_storage/errors/connection_errors.py
@@ -13,13 +13,8 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
-from cmdb.utils.error import CMDBError
 
-
-class DataBaseError(CMDBError):
-
-    def __init__(self, message: str = None):
-        self.message = message
+from cmdb.data_storage.errors import DataBaseError
 
 
 class DatabaseConnectionError(DataBaseError):
