+++ cmdb/data_storage/errors/database_errors.py
@@ -1,5 +1,5 @@
 # DATAGERRY - OpenSource Enterprise CMDB
-# Copyright (C) 2019 NETHINKS GmbH
+# Copyright (C) 2019 - 2020 NETHINKS GmbH
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -12,22 +12,4 @@
 # GNU Affero General Public License for more details.
 #
 # You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-import logging
-
-try:
-    from cmdb.utils.error import CMDBError
-except ImportError:
-    CMDBError = Exception
-
-LOGGER = logging.getLogger(__name__)
-
-
-class IDCounter:
-
-    COLLECTION = 'datastorage.counter'
-
-    def __init__(self, _id: str, counter: int):
-        self._id = _id
-        self.counter = counter
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
