+++ cmdb/exportd/exportd_job/exportd_job_base.py
@@ -57,7 +57,7 @@ class JobManagementBase:
         Returns:
             dict: json dump with database default encoding of the object attributes
         """
-        from cmdb.data_storage.database_utils import default
+        from cmdb.data_storage.utils import default
         import json
         return json.dumps(self.__dict__, default=default)
